requirements.txt
fastapi==0.124.2
uvicorn==0.38.0
sqlmodel==0.0.22
pydantic-settings==2.6.1
qdrant-client==1.16.2
mysql-connector-python==8.4.0
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
agno==2.3.12
google-genai==1.55.0
Pillow==11.0.0
python-multipart==0.0.18
